+++ sitevision/i18n/en.json
@@ -3,6 +3,7 @@
   "advanced": "Advanced",
   "user": "User",
   "assistant_id": "Assistant ID",
+  "group_chat": "Group chat",
   "app": "Application",
   "server_url": "Url to AI server",
   "user_info": "User information",
