+++ sitevision/i18n/sv.json
@@ -3,6 +3,7 @@
   "advanced": "Avancerat",
   "user": "Användare",
   "assistant_id": "Assistentens ID",
+  "group_chat": "Gruppchatt",
   "app": "Applikation",
   "server_url": "Url till AI-server",
   "user_info": "Användarinformation",
