+++ sitevision/manifest.json
@@ -1,6 +1,6 @@
 {
   "id": "sv-web-apps-service-assistant",
-  "version": "1.1.2",
+  "version": "1.2.0",
   "name": {
     "en": "Service assistant"
   },
