+++ sitevision/src/components/serverside-app/serverside-app.component.tsx
@@ -1,9 +1,7 @@
 import type { AssistantInfo } from "@sk-web-gui/ai";
 import * as React from "react";
-import {
-  AssistantDummie,
-  type Options,
-} from "../assistant-dummie/assistant-dummie.component";
+import { AssistantDummie } from "../assistant-dummie/assistant-dummie.component";
+import type { Options } from "../../types/options";
 
 export interface ServerSideAppProps {
   assistant: AssistantInfo;
