+++ sitevision/src/config/index.html
@@ -143,6 +143,14 @@
           <input class="form-control" name="assistantId" />
         </div>
         <div class="form-group">
+          <label><%= i18n('group_chat') %></label>
+          <input
+            type="checkbox"
+            data-value-type="boolean"
+            name="is_group_chat"
+          />
+        </div>
+        <div class="form-group">
           <label><%= i18n('app') %></label>
           <input class="form-control" name="app" />
         </div>
